Some changes and a question for you
Some changes, I added packages and change the beginning of the code a
little bit, now I will start with the linear models again so we have
everything settle.

Check one piece of code that looks like it doesn’t belong to any
function but it is in the function section. Check that when you have
time please!